Conflicts:
	upgrade.sh

+++ lib/MusicBrainz/Server/Controller/Edit/Relationship.pm
@@ -68,6 +68,9 @@ sub edit : Local RequireAuth Edit
         $c->detach;
     }
 
+    my $model0 = $c->model(type_to_model($type0));
+    my $model1 = $c->model(type_to_model($type1));
+
     $c->stash(
         root => $tree,
         type_info => JSON->new->latin1->encode(\%type_info),
@@ -110,7 +113,8 @@ sub edit : Local RequireAuth Edit
     my $form = $c->form(
         form => 'Relationship',
         init_object => $values,
-        attr_tree => $attr_tree
+        attr_tree => $attr_tree,
+        root => $tree
     );
     $form->field('link_type_id')->_load_options;
 
@@ -145,15 +149,19 @@ sub edit : Local RequireAuth Edit
             $c->detach;
         }
 
+        my $link_type = $c->model('LinkType')->get_by_id(
+            $form->field('link_type_id')->value
+        );
+
         my $values = $form->values;
         my $edit = $self->_insert_edit($c, $form,
             edit_type => $EDIT_RELATIONSHIP_EDIT,
             type0             => $type0,
             type1             => $type1,
-            entity0_id        => $form->field('entity0.id')->value,
-            entity1_id        => $form->field('entity1.id')->value,
+            entity0           => $model0->get_by_id($form->field('entity0.id')->value),
+            entity1           => $model1->get_by_id($form->field('entity1.id')->value),
             relationship      => $rel,
-            link_type_id      => $form->field('link_type_id')->value,
+            link_type         => $link_type,
             begin_date        => $form->field('begin_date')->value,
             end_date          => $form->field('end_date')->value,
             change_direction  => $form->field('direction')->value,
@@ -195,7 +203,7 @@ sub create : Local RequireAuth Edit
     my $source = $source_model->get_by_gid($source_gid);
     my $dest   = $dest_model->get_by_gid($dest_gid);
 
-    if ($source->id == $dest->id) {
+    if ($type0 eq $type1 && $source->id == $dest->id) {
         $c->stash( message => l('A relationship requires 2 different entities') );
         $c->detach('/error_500');
     }
@@ -222,7 +230,8 @@ sub create : Local RequireAuth Edit
 
     my $form = $c->form(
         form => 'Relationship',
-        attr_tree => $attr_tree
+        attr_tree => $attr_tree,
+        root => $tree
     );
     $c->stash(
         source => $source, source_type => $type0,
@@ -346,7 +355,11 @@ sub create_batch : Path('/edit/relationship/create-recordings') RequireAuth Edit
     my $attr_tree = $c->model('LinkAttributeType')->get_tree();
     $c->stash( attr_tree => $attr_tree );
 
-    my $form = $c->form( form => 'Relationship' );
+    my $form = $c->form(
+        form => 'Relationship',
+        attr_tree => $attr_tree,
+        root => $tree
+    );
     $c->stash(
         release => $release,
         dest    => $dest,
@@ -440,7 +453,11 @@ sub create_url : Local RequireAuth Edit
         type_info => JSON->new->latin1->encode(\%type_info),
     );
 
-    my $form = $c->form( form => 'Relationship::URL', reverse => $types[0] eq 'url' );
+    my $form = $c->form(
+        form => 'Relationship::URL',
+        reverse => $types[0] eq 'url',
+        root => $tree
+    );
 
     $c->stash(
         entity => $entity,
@@ -452,6 +469,18 @@ sub create_url : Local RequireAuth Edit
 
         my $e0 = $types[0] eq 'url' ? $url : $entity;
         my $e1 = $types[1] eq 'url' ? $url : $entity;
+
+        if ($c->model('Relationship')->exists(@types, {
+            link_type_id => $form->field('link_type_id')->value,
+            entity0 => $e0->id,
+            entity1 => $e1->id,
+        })) {
+            $c->stash(
+                exists => 1,
+                url => $form->field('url')->value
+            );
+            $c->detach;
+        }
 
         my $link_type = $c->model('LinkType')->get_by_id(
             $form->field('link_type_id')->value
